// Angular/project/src/app/modelos/login-registro.interface.ts
export interface loginRegistroI {
  ti_persona?: String | null | undefined;
  ni_persona?: Number | null | undefined;
  nombres?: String | null | undefined;
  apellidos?: String | null | undefined;
  email?: String | null | undefined;
  contrasena?: String | null | undefined;
  telefono?: Number | null | undefined;
  area?: String | null | undefined;
  rol?: Number | null | undefined;
}